refactor: remove UpdateExtractor dependency on SourceHandlerMixin
