+++ sqllineage/core/holders.py
@@ -1,5 +1,5 @@
 import itertools
-from typing import List, Optional, Set, Tuple, Union
+from typing import Dict, List, Optional, Set, Tuple, Union
 
 import networkx as nx
 from networkx import DiGraph
@@ -171,6 +171,27 @@ class SubQueryLineageHolder(ColumnLineageMixin):
                                     tgt_wildcard,
                                     src_wildcard,
                                 )
+
+    def get_alias_mapping_from_table_group(
+        self, table_group: List[Union[Path, Table, SubQuery]]
+    ) -> Dict[str, Union[Path, Table, SubQuery]]:
+        """
+        A table can be referred to as alias, table name, or database_name.table_name, create the mapping here.
+        For SubQuery, it's only alias then.
+        """
+        return {
+            **{
+                tgt: src
+                for src, tgt, attr in self.graph.edges(data=True)
+                if attr.get("type") == EdgeType.HAS_ALIAS and src in table_group
+            },
+            **{
+                table.raw_name: table
+                for table in table_group
+                if isinstance(table, Table)
+            },
+            **{str(table): table for table in table_group if isinstance(table, Table)},
+        }
 
     def _get_target_table(self) -> Optional[Union[SubQuery, Table]]:
         table = None
